Moves ExtraParams, GeoProp, and BathymetryResults to types.ts file. Renames getSketchToMpaProtectionLevel to getMpaProtectionLevel. Adds additional commenting.
