+++ src/components/Geomorphology.tsx
@@ -26,7 +26,7 @@ import {
 import project from "../../project";
 import Translator from "./TranslatorAsync";
 import { Trans, useTranslation } from "react-i18next";
-import { GeoProp } from "../clients/MpaTabReport";
+import { GeoProp } from "../util/types";
 
 const Number = new Intl.NumberFormat("en", { style: "decimal" });
