+++ src/components/ViabilityPage.tsx
@@ -4,9 +4,9 @@ import { SketchAttributesCard } from "@seasketch/geoprocessing/client-ui";
 import { OUSCard } from "./OUSCard";
 import { OusDemographics } from "./OusDemographic";
 import { OUSByIslandCard } from "./OUSByIslandCard";
-import { GeoProp } from "../clients/MpaTabReport";
 import { ProtectionCard } from "./ProtectionCard";
 import { BathymetryCard } from "./BathymetryCard";
+import { GeoProp } from "../util/types";
 
 const ReportPage: React.FunctionComponent<GeoProp> = (props) => {
   return (
