+++ src/functions/boundaryAreaOverlap.ts
@@ -25,11 +25,7 @@ import {
 } from "@seasketch/geoprocessing/client-core";
 import { getPrecalcMetrics } from "../../data/bin/getPrecalcMetrics";
 import { clipSketchToSubregion } from "../util/clipSketchToSubregion";
-
-interface ExtraParams {
-  /** Optional ID(s) of geographies to operate on. **/
-  geographies?: string[];
-}
+import { ExtraParams } from "../util/types";
 
 const metricGroup = project.getMetricGroup("boundaryAreaOverlap");
 const boundaryTotalMetrics = getPrecalcMetrics(
@@ -53,18 +49,15 @@ export async function boundaryAreaOverlap(
       includePercMetric: false,
     })
   ).map(
-    (metrics): Metric => ({
-      ...metrics,
+    (metric): Metric => ({
+      ...metric,
       classId: metricGroup.classes[0].classId,
     })
   );
 
   // Generate area metrics grouped by protection level, with area overlap within protection level removed
   // Each sketch gets one group metric for its protection level, while collection generates one for each protection level
-  const sketchToMpaClass = getSketchToMpaProtectionLevel(
-    clippedSketch,
-    areaMetrics
-  );
+  const sketchToMpaClass = getMpaProtectionLevel(clippedSketch);
   const metricToLevel = (sketchMetric: Metric) => {
     return sketchToMpaClass[sketchMetric.sketchId!];
   };
@@ -85,9 +78,13 @@ export async function boundaryAreaOverlap(
   };
 }
 
-export function getSketchToMpaProtectionLevel(
-  sketch: Sketch | SketchCollection | NullSketchCollection | NullSketch,
-  metrics: Metric[]
+/**
+ * Gets MPA Protection levels for all MPAs in a sketch collection from user attributes
+ * @param sketch User-created Sketch | SketchCollection
+ * @returns <string, string> mapping of sketchId to protection level
+ */
+export function getMpaProtectionLevel(
+  sketch: Sketch | SketchCollection | NullSketchCollection | NullSketch
 ): Record<string, string> {
   const sketchFeatures = getSketchFeatures(sketch);
   const protectionLevels = sketchFeatures.reduce<Record<string, string>>(
